Restaurants table with semantic UI
